feat/#38 기존 usecase Feature 로 몰아두기
