+++ Projects/Calendar/CalendarFeature/Sources/CalendarUseCase.swift
@@ -1,20 +1,13 @@
 //
-//  CalendarUseCaseProtocol.swift
-//  CalendarDomain
+//  CalendarUseCase.swift
+//  CalendarFeature
 //
 //  Created by 이지훈 on 7/3/25.
 //  Copyright © 2025 com.noweekend. All rights reserved.
 //
 
 import Foundation
-
-public protocol CalendarUseCaseProtocol {
-    func getCalendarEvents() async throws -> [CalendarEvent]
-    func getEventsForDate(_ date: Date) async throws -> [CalendarEvent]
-    func createCalendarEvent(_ event: CalendarEvent) async throws
-    func updateCalendarEvent(_ event: CalendarEvent) async throws
-    func deleteCalendarEvent(id: String) async throws
-}
+import CalendarDomain
 
 public class CalendarUseCase: CalendarUseCaseProtocol {
     private let calendarRepository: CalendarRepositoryProtocol
