+++ Projects/Profile/ProfileDomain/Sources/UseCases/UserUseCaseProtocol.swift
@@ -1,0 +1,14 @@
+//
+//  UserUseCaseProtocol.swift
+//  ProfileDomain
+//
+//  Created by 이지훈 on 7/3/25.
+//  Copyright © 2025 com.noweekend. All rights reserved.
+//
+
+import Foundation
+
+public protocol UserUseCaseProtocol {
+    func getCurrentUser() async throws -> User
+    func updateUser(_ user: User) async throws
+}
